Assignment1/requirements.txt
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.8.29
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
